Add manifestDiff algorithm
Given two build manifests, create an object with the added, changed and
removed files, between both manifests.

In order to provide atomic uploads, we need to figure out exactly which
files changed between deploys. Two achieve that we create and upload a
manifest.json file that describes the associated build.